requirements.txt
annotated-types==0.7.0
anyio==4.9.0
azure-ai-vision-imageanalysis==1.0.0
azure-cognitiveservices-speech==1.43.0
azure-core==1.34.0
certifi==2025.4.26
charset-normalizer==3.4.2
colorama==0.4.6
decorator==5.2.1
distro==1.9.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
imageio==2.37.0
imageio-ffmpeg==0.6.0
isodate==0.7.2
jiter==0.9.0
moviepy==2.1.2 
numpy==2.2.5
openai==1.77.0
pillow==10.4.0
proglog==0.1.11
pydantic==2.11.4
pydantic_core==2.33.2
PySide6==6.9.0
PySide6_Addons==6.9.0
PySide6_Essentials==6.9.0
python-dotenv==1.1.0
requests==2.32.3
shiboken6==6.9.0
six==1.17.0
sniffio==1.3.1
tqdm==4.67.1
typing-inspection==0.4.0
typing_extensions==4.13.2
urllib3==2.4.0
openai-whisper==20240930
python-Levenshtein